@@ -63,12 +63,21 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from 'vue-property-decorator';
+import { defineComponent } from 'vue'
 
-@Component
-export default class HelloWorld extends Vue {
-    @Prop() private msg!: string;
-}
+export default defineComponent({
+    name: 'HelloWorld',
+    props: {
+        msg: {
+            type: String,
+            required: false
+        }
+    },
+
+    setup(props) {
+        return {}
+    }
+})
 </script>
 
 <!-- Add "scoped" attribute to limit CSS to this component only -->
