@@ -1,18 +1,13 @@
-import Vue from 'vue';
-import Router from 'vue-router';
-import Home from './views/Home.vue';
+import { createRouter, createWebHistory } from 'vue-router'
+import Home from './views/Home.vue'
 
-// Router-Plugin wird verwendet
-Vue.use(Router);
-
-export default new Router({
-    mode: 'history',
-    base: process.env.BASE_URL,
+export default createRouter({
+    history: createWebHistory(process.env.BASE_URL),
     routes: [
         {
             name: 'home',
             path: '/',
-            component: Home,
+            component: Home
         },
         {
             name: 'about',
@@ -20,8 +15,7 @@ export default new Router({
             // route level code-splitting
             // this generates a separate chunk (about.[hash].js) for this route
             // which is lazy-loaded when the route is visited.
-            component: (): Promise<typeof import('*.vue')> =>
-                import(/* webpackChunkName: "about" */ './views/About.vue'),
-        },
-    ],
-});
+            component: () => import(/* webpackChunkName: "about" */ './views/About.vue')
+        }
+    ]
+})
