@@ -1,10 +1,8 @@
-import { CounterStore } from '../interfaces/CounterStore'
-
 type StoreProvider<T> = () => T
 
 export interface RootState {
     loaded: boolean
 
     // Sample...
-    counterStore: StoreProvider<CounterStore>
+    // counterStore: StoreProvider<CounterStore>
 }
